Beginning to draw out the concept of a resource
Since the FinAPI API follows a restful pattern a generalised approach
may work out:

The idea is that you have something like a session with the API which
can be queried for known resources. A resource is derived from the
method called on the session and matched in a dynamic manner, resulting
in a Resource object that is configured to call the endpoint. Let's see
where this path leads to...