BOOM! Took a wrong turn with EntityCollections
What seemed like a good idea proved to be over designed. Why? Because,
even though the tests pass, there is a lot wrong with this code:

I totally forgot to think about handling params in #all. Even
though this isn't necessary right now, the API has a lot of options for
filtering, sorting and the like that can be passed as parameters. This
will not work with the current implementation of EntityCollection.
EntityCollection has absolutely no clue about any parameters used to
receive the collection and will fail on subsequent pages.

Apart from the parameters, the enumeration has another problem: #all
doesn't return 'data' but a whole new EntityCollection which isn't expected
by the tests and currently just mocked out. This is a prime example of
the pitfalls of mocking! Not good, very bad coding, total failure...

I will try to keep thigs simple an just have #all return an Enumerator.
Let's see how this goes!